bid: refactored project
