+++ sources/bid/.build.mk
@@ -3,6 +3,8 @@ BID_SRC= $(wildcard sources/bid/*.c) \
 		 $(wildcard sources/bid/*/*/*.c) \
 		 $(wildcard sources/libraries/brutal/*.c) \
 		 $(wildcard sources/libraries/brutal/*/*.c) \
+		 $(wildcard sources/libraries/bid/*/*.c) \
+		 $(wildcard sources/libraries/bid/*.c) \
 		 $(wildcard sources/libraries/brutal/host/linux/*.c)
 
 BID_OBJ=$(patsubst sources/%.c, $(BUILDDIR_HOST)/%.c.o, $(BID_SRC))
