+++ sources/libraries/bid/bid.h
