+++ sources/libraries/bid/c_convert/bid_convert_c.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <brutal/base.h>
-#include <brutal/bid/bid.h>
+#include <bid/bid.h>
+#include <bid/parsing/bid_ast.h>
 #include <brutal/io.h>
 
 void convert_bid_to_c(const struct bid *bid, IoWriter *writer);
