+++ sources/libraries/bid/c_convert/bid_convert_method.c
@@ -1,67 +1,5 @@
-#include <brutal/alloc.h>
-#include <brutal/base.h>
-#include <brutal/bid/bid.h>
-#include <brutal/bid/bid_ast.h>
-#include <brutal/bid/bid_convert_c.h>
-#include <brutal/log.h>
-#include "brutal/text/str.h"
-
-static void bid_write_header(IoWriter *writer)
-{
-
-    print(writer, "#pragma once\n\n");
-    print(writer, "#include <brutal/base.h>\n\n");
-    print(writer, "// -- generated idl file --\n\n");
-}
-
-static Str bid_write_interface(IoWriter *writer, struct bid_ast_node *interface)
-{
-    print(writer, "// interface: {}\n\n", interface->interface.name);
-    return interface->interface.name;
-}
-
-static void bid_write_type(IoWriter *writer, struct bid_ast_node *type)
-{
-    if (type->children.length != 0)
-    {
-        print(writer, "{}(", type->ntype.name);
-        for (int i = 0; i < type->children.length; i++)
-        {
-            bid_write_type(writer, type->children.data[i]);
-            if (i + 1 < type->children.length)
-            {
-                print(writer, ", ");
-            }
-            else
-            {
-                print(writer, ")");
-            }
-        }
-    }
-    else
-    {
-        print(writer, "{}", type->ntype.name);
-    }
-}
-
-static void bid_write_typedef(IoWriter *writer, struct bid_ast_node *typedef_node)
-{
-    if (typedef_node->children.length < 2)
-    {
-        return;
-    }
-
-    struct bid_ast_node *type_node = typedef_node->children.data[1];
-    struct bid_ast_node *name_node = typedef_node->children.data[0];
-
-    print(writer, "typedef ");
-    bid_write_type(writer, type_node);
-
-    print(writer, " ");
-    bid_write_type(writer, name_node);
-
-    print(writer, "; \n\n");
-}
+#include <bid/c_convert/bid_convert_method.h>
+#include <bid/c_convert/bid_convert_type.h>
 
 void bid_write_argument(IoWriter *writer, struct bid_ast_node *arg)
 {
@@ -96,7 +34,6 @@ void bid_write_method_received_structure(IoWriter *writer, struct bid_ast_node *
     {
         if (method->children.data[i]->type == BID_AST_NODE_TYPE_METHOD_RETURN_TYPE)
         {
-            log("writing bid_write_method_received_structure {}/{}", i, method->children.length);
             bid_write_argument(writer, method->children.data[i]);
             print(writer, "; \n");
         }
@@ -120,7 +57,7 @@ static void bid_write_method_send_structure(IoWriter *writer, struct bid_ast_nod
     print(writer, "}; \n\n");
 }
 
-static void bid_write_method(IoWriter *writer, struct bid_ast_node *method, Str current_namespace)
+void bid_write_method(IoWriter *writer, struct bid_ast_node *method, Str current_namespace)
 {
 
     print(writer, "// - {}.{} \n\n", current_namespace, method->method.name);
@@ -164,34 +101,4 @@ static void bid_write_method(IoWriter *writer, struct bid_ast_node *method, Str 
     }
     */
     print(writer, ");\n\n");
-}
-
-static void bid_write_node(IoWriter *writer, struct bid_ast_node *node, Str current_namespace)
-{
-    switch (node->type)
-    {
-    case BID_AST_NODE_TYPE_INTERFACE:
-        current_namespace = bid_write_interface(writer, node);
-        for (int i = 0; i < node->children.length; i++)
-        {
-            bid_write_node(writer, node->children.data[i], current_namespace);
-        }
-        break;
-    case BID_AST_NODE_TYPE_METHOD:
-        bid_write_method(writer, node, current_namespace);
-        break;
-    case BID_AST_NODE_TYPE_TYPEDEF:
-        bid_write_typedef(writer, node);
-        break;
-
-    default:
-        break;
-    }
-}
-
-void convert_bid_to_c(const struct bid *bid, IoWriter *writer)
-{
-    bid_write_header(writer);
-
-    bid_write_node(writer, bid->root_ast, str_cast(""));
 }
