+++ sources/libraries/bid/parsing/bid_interface.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <brutal/base.h>
-#include <brutal/bid/bid_ast.h>
+#include <bid/parsing/bid_ast.h>
 
 BidParseResult scan_interface(struct bid *idl_in, struct bid_ast_node *type_node);
