+++ sources/libraries/bid/parsing/bid_type.h
@@ -1,5 +1,5 @@
 #pragma once
 #include <brutal/base.h>
-#include <brutal/bid/bid_ast.h>
+#include <bid/parsing/bid_ast.h>
 
 BidParseResult scan_type(struct bid *idl_in, struct bid_ast_node *type_node);
